python_server/requirements.txt
# AI Avatar Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
orjson==3.9.10

# AI/ML Dependencies  
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Audio Dependencies
elevenlabs==0.2.26
azure-cognitiveservices-speech==1.34.1

# Database Dependencies
sqlalchemy==2.0.23
alembic==1.13.1
psycopg==3.1.13

# Text Processing
janome==0.5.0
jaconv==0.3.4
pandas==2.1.4

# PDF Processing
pdf2image==1.17.0
Pillow==10.1.0

# Utility Dependencies
structlog==23.2.0
python-multipart==0.0.6